+++ src/main/java/com/ddevuss/weather/oracle/entity/Location.java
@@ -6,19 +6,17 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Entity
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 @Table(name = "locations")
-public class Location implements BaseEntity<Integer> {
+public class Location implements BaseEntity<Long> {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String name;
 
@@ -26,7 +24,7 @@ public class Location implements BaseEntity<Integer> {
     @JoinColumn(name = "user_id")
     private User user;
 
-    private BigDecimal latitude;
+    private Double latitude;
 
-    private BigDecimal longitude;
+    private Double longitude;
 }
